4.21 done

+++ part4/blog/controllers/blogs.js
@@ -42,8 +42,23 @@ blogsRouter.post("/", async (request, response) => {
     response.status(201).json(savedBlog.toJSON())
 })
 blogsRouter.delete("/:id",async (request,response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: "token invalid" })
+    }
+    const id = await request.params.id
+    const blog = await Blog.findById(request.params.id)
+
+    if ( blog.user.toString() === decodedToken.id.toString() ){
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    }
+    else {
+        response.status(401).json({
+            error: "missing premission to do this"})
+
+    }
+
 
 })
 blogsRouter.put("/:id", async (request, response) => {
